MAINT: move optional into `utils` folder
