+++ spectrochempy/core/dataset/nddataset.py
@@ -30,6 +30,7 @@ from traitlets import (
     validate,
 )
 from traittypes import Array
+from utils.optional import import_optional_dependency
 
 from spectrochempy.core import error_, warning_
 from spectrochempy.core.dataset.coord import Coord, LinearCoord
@@ -40,7 +41,6 @@ from spectrochempy.core.dataset.ndio import NDIO
 from spectrochempy.core.dataset.ndmath import NDMath, _set_operators, _set_ufuncs
 from spectrochempy.core.dataset.ndplot import NDPlot
 from spectrochempy.core.project.baseproject import AbstractProject
-from spectrochempy.optional import import_optional_dependency
 from spectrochempy.utils import colored_output, get_user_and_node
 from spectrochempy.utils.exceptions import SpectroChemPyError, UnknownTimeZoneError
