+++ spectrochempy/core/readers/download.py
@@ -16,12 +16,12 @@ from pathlib import Path
 
 import numpy as np
 import requests
+from utils.optional import import_optional_dependency
 
 from spectrochempy.core import error_, info_
 from spectrochempy.core.dataset.coord import Coord
 from spectrochempy.core.dataset.nddataset import NDDataset
 from spectrochempy.core.readers.read_jcamp import read_jcamp
-from spectrochempy.optional import import_optional_dependency
 from spectrochempy.utils import is_iterable
