+++ spectrochempy/utils/optional.py
