+++ spectrochempy/utils/plots.py
@@ -11,10 +11,10 @@ import matplotlib.axes as maxes
 import mpl_toolkits.mplot3d.axes3d as maxes3D  # noqa: N812
 import numpy as np
 from matplotlib import pyplot as plt
+from utils.optional import import_optional_dependency
 
 from spectrochempy.core.dataset.meta import Meta
 from spectrochempy.core.units import remove_args_units
-from spectrochempy.optional import import_optional_dependency
 
 __all__ = [
     "cmyk2rgb",
